ZMHandlersTest: fix up multiline filepaths
The same zmmultipoints.shp was used for all three test cases. This meant
the mmultipoints.shp and zmultipoints.shp files were completely unused
by the tests.

As a flow-on effect, the m value in the testReadMMultiPoints() needs to
be fixed up to match the file. shpdump shows the first point as being:
  (1208.59837545126,924.85559566787, 0, 20)