ZMHandlersTest: small refactor to avoid repeated code
The testReadFlatXYZ functions all repeated the same code. Refactor it
out into a helper function. I've tweaked it to return the number of
features read, which results in a fairly straight-forward assertion.

There really should be tests that we can read all the features for all
the various Z/M file permutations - this change will help with that.